+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -15,7 +15,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
-import kitchenpos.external.profanity.ProfanityClient;
+import kitchenpos.profanity.ProfanityClient;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuProduct;
